requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
chromadb==0.4.22
google-generativeai==0.3.2
pydantic==2.5.3
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.6
